+++ packages/graphql-server/package.json
@@ -67,7 +67,6 @@
     "koa-bodyparser": "^4.2.1",
     "koa-morgan": "^1.0.1",
     "koa-mount": "^4.0.0",
-    "koa-proxies": "^0.11.0",
     "koa-router": "^7.4.0",
     "koa-static": "^5.0.0",
     "koa-views": "^6.1.4",
